Fix nasty degenerative remove hashmap case

+++ common.h
@@ -8,7 +8,7 @@
 #include <stdbool.h>
 #include "time.h"
 
-typedef uint64_t u64;
+typedef long long unsigned u64;
 typedef uint32_t u32;
 typedef uint16_t u16;
 typedef uint8_t u8;
